Extract fixtures for Validation tests.
