+++ tests/Validation/Fixtures/EloquentModelStub.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Illuminate\Tests\Validation\Fixtures;
+
+use Illuminate\Database\Eloquent\Model;
+
+class EloquentModelStub extends Model
+{
+    protected $primaryKey = 'id_column';
+
+    protected $guarded = [];
+}
