+++ tests/Validation/Fixtures/EloquentTestUser.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Illuminate\Tests\Validation\Fixtures;
+
+use Illuminate\Database\Eloquent\Model as Eloquent;
+
+class EloquentTestUser extends Eloquent
+{
+    protected $table = 'users';
+
+    protected $guarded = [];
+
+    public $timestamps = false;
+}
