+++ tests/Validation/Fixtures/image.png
