+++ tests/Validation/ValidationFactoryTest.php
@@ -13,6 +13,8 @@ class ValidationFactoryTest extends TestCase
 {
     public function tearDown()
     {
+        parent::tearDown();
+
         m::close();
     }
